Move code to source directory
This is the first commit in a series of commits that will
attempt to bring the structure of this challenge inline
with other challenges.

Based on student experiences with this challenge, the
in-memory database and single file structure will be
removed.

This commit adds a source directory where all code will be added.